Moved RDMPCollectionCommonFunctionality to its own branch
